requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.3.29
click==8.1.7
clickhouse-connect==0.7.6
coloredlogs==15.0.1
dataclasses-json==0.6.4
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
docx2txt==0.8
duckdb==0.10.1
exceptiongroup==1.2.0
fastapi==0.85.1
filelock==3.13.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
glob2==0.7
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.0
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.0
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.13
langchain-anthropic==0.1.4
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.31
lz4==4.3.3
marshmallow==3.21.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.14.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
posthog==3.5.0
protobuf==4.25.3
psycopg2==2.9.9
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydantic_core==2.16.3
PyPika==0.48.9
pyproject_hooks==1.0.0
pysqlite3-binary==0.5.2.post3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
starlette==0.20.4
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
typer==0.10.0
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
whitenoise==6.6.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
